app/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
boto3==1.33.11
botocore==1.33.11
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-limiter==0.1.5
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
jmespath==1.0.1
limits==3.7.0
loguru==0.7.2
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
redis==4.6.0
s3transfer==0.8.2
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
win32-setctime==1.1.0
wrapt==1.16.0
zipp==3.17.0
